This is a set of edits related to #123 and #125. These need to be vetted
This:
 * SF298 is taller, allowing the distributions statement and abstract boxes to be larger
 * whitespace on cover has been re-worked to accomodate long distribution statements
 * 'compacttile' option is introduced, which gives a compact / two column version of the signature lines on the title page (iii)

One 'trick' here is to put the distribution statement into a minipage, measure its height and if its
too tall, to switch to a smaller font, in a best-effort attempt to accomodate long statements

There are, of course, still scenarios where we can't fit all the data, like
 - joint thesis (two authors)
 - coadvisors, second reader
 - dual-degree (two degrees, two department chairs)

In general 1pc and 2pc have been changed to the equiv. 6pt and 12pt,
and vfill changed to the equivalent vspace{0 plus 1fill}, then
each of these were modified to support flexing "down" with 'minus 5pt' or
similar logic.